backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
email-validator==2.1.0
sqlalchemy==2.0.31
psycopg[binary]==3.1.19
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.8
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
orjson==3.10.3
itsdangerous==2.2.0
numpy==1.26.4
Pillow==10.3.0
openpyxl==3.1.5
reportlab==4.4.6
pgvector==0.2.5
insightface==0.7.3
onnx==1.16.0
onnxruntime==1.19.2
opencv-python==4.9.0.80
ml-dtypes==0.3.2
google-generativeai==0.7.2
cryptography==42.0.5
qrcode[pil]==7.4.2
python-magic==0.4.27
boto3==1.34.90
celery[redis]==5.4.0
scikit-learn==1.4.2
tensorflow==2.16.1
icalendar==5.0.11

# Development tools
black==24.8.0
ruff==0.6.3
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-cov==5.0.0
